+++ Source/Core/Visualization/Renderer/PolygonRendererGLSL.h
@@ -11,6 +11,8 @@
 #include <kvs/Shader>
 #include <kvs/ProgramObject>
 #include <kvs/VertexBufferObjectManager>
+#include <kvs/Deprecated>
+#include <string>
 
 
 namespace kvs
@@ -30,13 +32,13 @@ class PolygonRenderer : public kvs::PolygonRenderer
     kvsModuleBaseClass( kvs::PolygonRenderer );
 
 private:
+    std::string m_vert_file; ///< vertex shader file
+    std::string m_frag_file; ///< fragment shader file
     size_t m_width; ///< window width
     size_t m_height; ///< window height
     const kvs::ObjectBase* m_object; ///< pointer to the rendering object
-    bool m_has_normal; ///< check flag for the normal array
-    bool m_has_connection; ///< check flag for the connection array
     float m_polygon_offset; ///< polygon offset
-    kvs::Shader::ShadingModel* m_shader; ///< shading method
+    kvs::Shader::ShadingModel* m_shading_model; ///< shading method
     kvs::ProgramObject m_shader_program; ///< shader program
     kvs::VertexBufferObjectManager m_vbo_manager; ///< vertex buffer object manager
 
@@ -46,31 +48,51 @@ public:
 
     void exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Light* light );
 
+    const std::string& vertexShaderFile() const { return m_vert_file; }
+    const std::string& fragmentShaderFile() const { return m_frag_file; }
     float polygonOffset() const { return m_polygon_offset; }
 
-    template <typename ShadingType>
-    void setShader( const ShadingType shader );
+    void setVertexShaderFile( const std::string& vert_file ) { m_vert_file = vert_file; }
+    void setFragmentShaderFile( const std::string& frag_file ) { m_frag_file = frag_file; }
     void setPolygonOffset( const float offset ) { m_polygon_offset = offset; }
 
-private:
-    void create_shader_program();
-    void create_buffer_object( const kvs::PolygonObject* point );
-};
-
-template <typename ShadingType>
-inline void PolygonRenderer::setShader( const ShadingType shader )
-{
-    if ( m_shader )
+    void setShaderFiles( const std::string& vert_file, const std::string& frag_file )
     {
-        delete m_shader;
-        m_shader = NULL;
+        this->setVertexShaderFile( vert_file );
+        this->setFragmentShaderFile( frag_file );
     }
 
-    m_shader = new ShadingType( shader );
-    if ( !m_shader )
+    template <typename Model>
+    void setShadingModel( const Model model )
     {
-        kvsMessageError("Cannot create a specified shader.");
+        if ( m_shading_model ) { delete m_shading_model; m_shading_model = NULL; }
+        m_shading_model = new Model( model );
+        if ( !m_shading_model )
+        {
+            kvsMessageError("Cannot create a specified shading model.");
+        }
     }
+
+    template <typename ShadingType>
+    KVS_DEPRECATED( void setShader( const ShadingType shader ) ) { this->setShadingModel<ShadingType>( shader ); }
+
+protected:
+    kvs::Shader::ShadingModel& shadingModel() { return *m_shading_model; }
+    kvs::ProgramObject& shader() { return m_shader_program; }
+    kvs::VertexBufferObjectManager& vboManager() { return m_vbo_manager; }
+
+    bool isWindowCreated() { return m_width == 0 && m_height == 0; }
+    bool isWindowResized( size_t w, size_t h ) { return m_width != w || m_height != h; }
+    bool isObjectChanged( const kvs::ObjectBase* o ) { return m_object != o; }
+    void setWindowSize( size_t w, size_t h ) { m_width = w; m_height = h; }
+    void attachObject( const kvs::ObjectBase* o ) { m_object = o; }
+
+    void createShaderProgram();
+    void updateShaderProgram();
+    void setupShaderProgram();
+    void createBufferObject( const kvs::PolygonObject* polygon );
+    void updateBufferObject( const kvs::PolygonObject* polygon );
+    void drawBufferObject( const kvs::PolygonObject* polygon );
 };
 
 } // end of namespace glsl